fixed the font loading issue
the problem was you were trying to make the public folder together with images folder as your project root, but you also try to access the fonts at the fonts folder which is not included as project root. I changed your project root to /, so everything becomes accessible. If you prefer, another solution is to make the fonts folder also part of the root